Safari-609.1.11.1/LayoutTests/svg/animations/animate-dynamic-update-attributeName-expected.txt

SVG 1.1 dynamic animation tests

Test behavior on dynamic-update of attributeName

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS colorComponents[1] is 255
PASS colorComponents[2] is 0
PASS colorComponents[3] is 0
PASS rect.style.color is ""
PASS colorComponents[1] is 0
PASS colorComponents[2] is 128
PASS colorComponents[3] is 0
PASS rect.style.color is ""
PASS colorComponents[1] is 0
PASS colorComponents[2] is 128
PASS colorComponents[3] is 0
PASS rect.style.color is ""
PASS rect.style.fill is ""
PASS successfullyParsed is true

TEST COMPLETE
